+++ src/transforms/v2-to-v3/__fixtures__/api-promise/client-class-member.input.ts
@@ -1,0 +1,18 @@
+import AWS from "aws-sdk";
+
+// Client as class member
+class ClientClassMember {
+  private clientInClass: AWS.DynamoDB;
+  
+  constructor(clientInCtr: AWS.DynamoDB) {
+    this.clientInClass = clientInCtr;
+  }
+
+  async listTables() {
+    return this.clientInClass.listTables().promise();
+  }
+
+  async listTagsOfResource() {
+    return this.clientInClass.listTagsOfResource({ ResourceArn: "STRING_VALUE" }).promise();
+  }
+}
