+++ src/transforms/v2-to-v3/__fixtures__/api-promise/client-class-member.output.ts
@@ -1,0 +1,18 @@
+import { DynamoDB } from "@aws-sdk/client-dynamodb";
+
+// Client as class member
+class ClientClassMember {
+  private clientInClass: DynamoDB;
+  
+  constructor(clientInCtr: DynamoDB) {
+    this.clientInClass = clientInCtr;
+  }
+
+  async listTables() {
+    return this.clientInClass.listTables();
+  }
+
+  async listTagsOfResource() {
+    return this.clientInClass.listTagsOfResource({ ResourceArn: "STRING_VALUE" });
+  }
+}
